HUGE REFACTOR : abilities, regeneration (to be tested), UniqueIDProviders, loaders (UIDP also), View (GUI), something else ; missing: tests, refactoring Equip Upgrade values to new levels and Life/Mana Max standars

+++ JavaBackup/AllProjects/src/dataStructures/isom/InSpaceObjectsManager.java
@@ -92,7 +92,7 @@ public interface InSpaceObjectsManager<Distance extends Number>
 
 	// TODO OTHER
 
-	public abstract ObjectLocated getObjectLocated(Integer ID);
+	public abstract ObjectLocated getObjectLocated(Long ID);
 
 	public default boolean removeAllObjects() {
 		clearAllNodes();
